bot/requirements.txt
aiogram==3.3.0
yt-dlp==2023.11.16
ffmpeg-python==0.2.0
python-dotenv==1.0.0